Useless test removed
